requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-pytest==0.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.0.2
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyaml==20.4.0
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
